+++ packages/menu/src/MenuContext/MenuContext.tsx
@@ -1,8 +1,24 @@
 import { createContext, useContext } from 'react';
 
 import { createDescendantsContext } from '@leafygreen-ui/descendants';
+import { Descendant } from '@leafygreen-ui/descendants';
+import { Theme } from '@leafygreen-ui/lib';
 
-import { MenuContextData } from './MenuContext.types';
+import { HighlightReducerReturnType } from '../HighlightReducer/highlight.types';
+
+export interface MenuContextData {
+  theme: Theme;
+  darkMode: boolean;
+
+  /** The index of the currently highlighted (focused) item */
+  highlight?: Descendant;
+
+  /** Sets the current highlight by index or id */
+  setHighlight?: HighlightReducerReturnType['setHighlight'];
+
+  /** Whether to a dark menu in light mode */
+  renderDarkMenu?: boolean;
+}
 
 export const MenuDescendantsContext = createDescendantsContext(
   'MenuDescendantsContext',
@@ -15,5 +31,3 @@ export const MenuContext = createContext<MenuContextData>({
 });
 
 export const useMenuContext = () => useContext(MenuContext);
-
-export default MenuContext;
