+++ packages/menu/src/MenuContext/index.ts
@@ -1,5 +1,12 @@
 export {
-  default as MenuContext,
+  MenuContext,
+  type MenuContextData,
   MenuDescendantsContext,
   useMenuContext,
 } from './MenuContext';
+export {
+  SubMenuContext,
+  type SubMenuContextData,
+  SubMenuProvider,
+  useSubMenuContext,
+} from './SubMenuContext';
