+++ packages/menu/src/SubMenu/index.ts
@@ -1,3 +1,2 @@
 export { SubMenu } from './SubMenu';
 export { InternalSubMenuProps, SubMenuProps } from './SubMenu.types';
-export { useSubMenuContext } from './SubMenuContext';
